+++ src/main/java/io/github/forgestove/create_cyber_goggles/mixin/chainConveyor/ChainConveyorRidingHandlerMixin.java
@@ -39,6 +39,6 @@ public abstract class ChainConveyorRidingHandlerMixin {
 			ChainConveyorRidingHandler.ridingChainConveyor,
 			false
 		));
-		if (Common.testForStealth(player)) player.connection.send(new ServerboundPlayerCommandPacket(player, Action.PRESS_SHIFT_KEY));
+		if (CCGHelper.testForStealth(player)) player.connection.send(new ServerboundPlayerCommandPacket(player, Action.PRESS_SHIFT_KEY));
 	}
 }